requirements.txt
langchain
langgraph
langchain-community
langchain-text-splitters
faiss-cpu
sentence-transformers
langchain-google-genai